requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
rich==13.7.0
requests==2.31.0
openai==1.64.0
flask-caching==2.1.0
